requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-sympycharfield==0.4.1
djangorestframework==3.13.1
mpmath==1.2.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
sympy==1.10.1
